update bunny instrumentation
